Add missing '-input /dev/null' to rofi when asking for password (#9)
When refactoring this got removed. Which hangs the input when running
via gpg. It doesn't show up when running `pinentry-rofi` directly. The
hang is pretty nasty as well, as input to other programs stop working.
Only way is to open up a new terminal and find the PID and kill it.