Add the rest of the standard cli options as dummy options [#29]
Not sure how most of these will translate to pinentry-rofi.  Timeout
is the only one that make sense to implement.  But I'm not sure how
the best way would be to implement it.  Rofi does have a timeout
configuration, but you need to set that in the config file and not on
the cli.

So for now I'm just going to expose these as dummy options to not
error out on tools that expect these options to be there.