Uri encode password before sending it to gnupg [#27]
It will strip away any trailing newlines before encoding the password
and append one when formatting the string for gpg.  That to avoid
encoding any trailing newline which gpg expects in the answer.

This fixes a major bug in the way pinentry-rofi was giving the
password to gnupg.  Gnupg expected the password to be uri encoded,
which it was not.  With this commit the password is now uri encoded
before handing it over to gnupg.

This means passwords containing special characters that was created
with pinentry-rofi before this commit _will not work_ as expected
going forward.  Some user intervention will be needed when switching
to this.

This should fix #27.